+++ tests/python/relax/test_from_exported_to_cuda.py
@@ -754,5 +754,25 @@ def test_mul(target, dev):
     assert_torch_output_vs_tvm_from_exported_to_cuda(raw_data, torch_module, target, dev)
 
 
+@tvm.testing.parametrize_targets("cuda")
+def test_concat(target, dev):
+    class ConcatFour(nn.Module):
+        def __init__(self, dim=0):
+            super(ConcatFour, self).__init__()
+            self.dim = dim
+            self.x2 = torch.randn(2, 3)
+            self.x3 = torch.randn(2, 3)
+            self.x4 = torch.randn(2, 3)
+
+        def forward(self, x):
+            return torch.cat((x ,self.x2, self.x3, self.x4), dim=self.dim)
+            
+    torch_module = ConcatFour().eval()
+
+    raw_data = np.random.rand(2,3).astype("float32")
+
+    assert_torch_output_vs_tvm_from_exported_to_cuda(raw_data, torch_module, target, dev)
+
+
 if __name__ == "__main__":
     tvm.testing.main()
